Improved SW metrics.SW and SWG swapped
Added java doc.
Swapped SW and SWG. Implementations were reversed originally.
